questions: Crear módulo utils
